+++ src/lib/pages/deploy/index.tsx
@@ -1,13 +1,12 @@
 import { Heading, Text } from "@chakra-ui/react";
-import { useRouter } from "next/router";
 
+import { useInternalNavigate } from "lib/app-provider";
+import { ButtonCard } from "lib/components/ButtonCard";
 import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
 
-import { ButtonCard } from "./components/ButtonCard";
-
 const Deploy = () => {
-  const router = useRouter();
+  const navigate = useInternalNavigate();
   return (
     <WasmPageContainer>
       <Text variant="body1" color="text.dark" mb={3} fontWeight={700}>
@@ -20,13 +19,13 @@ const Deploy = () => {
       <ButtonCard
         title="Upload new WASM File"
         description="Store a new Wasm file on-chain"
-        onClick={() => router.push("/upload")}
+        onClick={() => navigate({ pathname: "/upload" })}
         mb="16px"
       />
       <ButtonCard
         title="Use existing Code IDs"
         description="Input code ID or select from previously stored codes or your saved codes"
-        onClick={() => router.push("/instantiate")}
+        onClick={() => navigate({ pathname: "/instantiate" })}
       />
     </WasmPageContainer>
   );