+++ src/lib/pages/past-txs/components/PastTxTable.tsx
@@ -17,6 +17,7 @@ import type { MouseEvent } from "react";
 import { BsArrowCounterclockwise } from "react-icons/bs";
 import { MdCheck, MdClose, MdKeyboardArrowDown } from "react-icons/md";
 
+import { useRedo } from "../hooks/useRedo";
 import { useFabricateFee, useSimulateFee } from "lib/app-provider";
 import { useResendTx } from "lib/app-provider/tx/resend";
 import type { SingleMsgProps } from "lib/components/action-msg/SingleMsg";
@@ -40,7 +41,6 @@ import {
   formatUDenom,
   formatUToken,
   extractMsgType,
-  onClickRedo,
 } from "lib/utils";
 
 import { MsgDetail } from "./MsgDetail";
@@ -52,6 +52,7 @@ interface PastTxTableProps {
 }
 
 const PastTxTable = ({ element }: PastTxTableProps) => {
+  const onClickRedo = useRedo();
   const { isOpen, onToggle } = useDisclosure();
   const [isAccordion, setIsAccordion] = useState(false);
   const [button, setButton] = useState<"redo" | "resend" | "">("");
@@ -59,7 +60,7 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
   const [error, setError] = useState("");
   const { currentChainName } = useWallet();
 
-  const { getContractInfo } = useContractStore();
+  const { getContractLocalInfo } = useContractStore();
 
   const extractMessage = useCallback((data: Transaction) => {
     const uploadMsgs: DetailUpload[] = [];
@@ -169,7 +170,9 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
         );
       }
       setButton("redo");
-      const contractInfo = getContractInfo(instantiateMsgs[0].contractAddress);
+      const contractLocalInfo = getContractLocalInfo(
+        instantiateMsgs[0].contractAddress
+      );
       // Only 1 Instantiate Msgs
       const singleMsgProps: SingleMsgProps = element.success
         ? {
@@ -177,7 +180,8 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
             text1: "contract",
             link1: {
               type: "contract_address",
-              value: contractInfo?.name || instantiateMsgs[0].contractAddress,
+              value:
+                contractLocalInfo?.name || instantiateMsgs[0].contractAddress,
               copyValue: instantiateMsgs[0].contractAddress,
             },
             text3: "from Code ID",
@@ -196,7 +200,7 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
           };
       return <SingleMsg {...singleMsgProps} />;
     },
-    [element.success, getContractInfo]
+    [element.success, getContractLocalInfo]
   );
 
   // TODO - Refactor
@@ -237,7 +241,7 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
         setIsAccordion(false);
       }
 
-      const contractInfo = getContractInfo(executeMsgs[0].contract);
+      const contractLocalInfo = getContractLocalInfo(executeMsgs[0].contract);
 
       // Only 1 Execute Msg
       const singleMsgProps: SingleMsgProps = element.success
@@ -248,7 +252,7 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
             text2: "on",
             link1: {
               type: "contract_address",
-              value: contractInfo?.name || executeMsgs[0].contract,
+              value: contractLocalInfo?.name || executeMsgs[0].contract,
               copyValue: executeMsgs[0].contract,
             },
           }
@@ -257,13 +261,13 @@ const PastTxTable = ({ element }: PastTxTableProps) => {
             text1: "to execute message from",
             link1: {
               type: "contract_address",
-              value: contractInfo?.name || executeMsgs[0].contract,
+              value: contractLocalInfo?.name || executeMsgs[0].contract,
               copyValue: executeMsgs[0].contract,
             },
           };
       return <SingleMsg {...singleMsgProps} />;
     },
-    [element.success, getContractInfo]
+    [element.success, getContractLocalInfo]
   );
 
   // TODO - Refactor