+++ src/lib/styles/theme/components/form.ts
@@ -1,32 +1,5 @@
 import type { ComponentStyleConfig } from "@chakra-ui/react";
 
-const activeLabelStyles = {
-  "&.md-form, &.textarea-form": {
-    _focusWithin: {
-      label: {
-        transform: "scale(0.75) translateY(-20px)",
-        lineHeight: "1.2",
-      },
-    },
-    "& .floating": {
-      transform: "scale(0.75) translateY(-20px)",
-      lineHeight: "1.2",
-    },
-  },
-  "&.lg-form": {
-    _focusWithin: {
-      label: {
-        lineHeight: "1.2",
-        transform: "scale(0.75) translateY(-28px)",
-      },
-    },
-    "& .floating": {
-      transform: "scale(0.75) translateY(-28px)",
-      lineHeight: "1.2",
-    },
-  },
-};
-
 export const Form: ComponentStyleConfig = {
   baseStyle: {
     container: {
@@ -38,7 +11,6 @@ export const Form: ComponentStyleConfig = {
   variants: {
     floating: {
       container: {
-        ...activeLabelStyles,
         label: {
           top: 0,
           left: 0,
@@ -47,6 +19,7 @@ export const Form: ComponentStyleConfig = {
           px: 1,
           position: "absolute",
           fontWeight: "400",
+          lineHeight: "1.2",
           pointerEvents: "none",
           transformOrigin: "left top",
           color: "text.dark",
@@ -55,6 +28,12 @@ export const Form: ComponentStyleConfig = {
           },
           _active: {
             color: "primary.main",
+          },
+          "&.md-label, &.textarea-label": {
+            transform: "scale(0.75) translateY(-20px)",
+          },
+          "&.lg-label": {
+            transform: "scale(0.75) translateY(-28px)",
           },
         },
         "div.helper-text, .error-text": {