+++ src/lib/utils/base64.ts
@@ -1,4 +1,5 @@
 export const decode = (str: string): string =>
   Buffer.from(str, "base64").toString("binary");
+
 export const encode = (str: string): string =>
   Buffer.from(str, "binary").toString("base64");