+++ src/pages/[network]/instantiate.tsx
@@ -1,0 +1,3 @@
+import Instantiate from "lib/pages/instantiate";
+
+export default Instantiate;